refactor: migrate details grid
